migrated dump data from sqlite3 to postgreSQL

+++ Orders/models.py
@@ -58,7 +58,7 @@ class Order(models.Model):
      address=models.CharField(max_length=25)
      city=models.CharField(max_length=25)
      postal_code=models.CharField(max_length=25)
-     contact=models.PositiveIntegerField(null=True,blank=True)
+     contact=models.BigIntegerField(null=True,blank=True)
      country=models.CharField(max_length=100,default='Nepal')
      total_amount=models.DecimalField( max_digits=10, decimal_places=2)
      tax=models.DecimalField( max_digits=10, decimal_places=2, default=0.00)
